extract functions from api route
